% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemix_drift.R
\name{custom_convert}
\alias{custom_convert}
\title{internal function}
\usage{
custom_convert(mat)
}
\arguments{
\item{mat}{the pairwise matrix in matrix class}
}
\value{
a long tibble
}
\description{
convert the pairwise matrix into long table that can be plot by geom_tile
}
\keyword{internal}
