% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_treemix.R
\docType{data}
\name{obj_treemix}
\alias{obj_treemix}
\title{main obj from package}
\format{
a list with multiple elements
}
\usage{
data(obj_treemix)
}
\description{
example result from a treemix run on msprime simulation data with 1 migration
edge and block 10000
}
\keyword{datasets}
