% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_treemixResult.R
\name{parse_verDf_fromObj}
\alias{parse_verDf_fromObj}
\title{Parse vertices table from treemix object}
\usage{
parse_verDf_fromObj(obj, ...)
}
\arguments{
\item{obj}{object from read_treemixResult function}
}
\value{
a named data frame of df.vertices
}
\description{
this is only a reminder with named data frame
}
\examples{
parse_verDf_fromObj(obj_treemix)
}
