% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemix_graph.R
\name{plot_treemix_graph}
\alias{plot_treemix_graph}
\title{Plot a treemix graph}
\usage{
plot_treemix_graph(obj)
}
\arguments{
\item{obj}{an object of class treemixResult}
}
\value{
a ggplot object
}
\description{
Plot a treemix graph with ggplot from the treemixResult obj
}
\examples{
infiles <- system.file('extdata', package='treemixTools') |> list.files(full.names=TRUE)
inStem <- infiles[1] |> gsub(pattern=".cov.gz", replacement="")
obj <- read_treemixResult(inStem)
plot_treemix_graph(obj)

}
