% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_treemixResult.R
\name{read_treemixResult}
\alias{read_treemixResult}
\title{Read treemix result}
\usage{
read_treemixResult(stem, ...)
}
\arguments{
\item{inStem}{Stem of the path to the treemix result}
}
\value{
a list of treemix result
}
\description{
Check if files exist and read treemix result into a list
}
\examples{
infiles <- system.file('extdata', package='treemixTools') |> list.files(full.names=TRUE)
inStem <- infiles[1] |> gsub(pattern=".cov.gz", replacement="")
obj <- read_treemixResult(inStem)
}
