% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemix_plot_graph.R
\name{treemix_plot_graph}
\alias{treemix_plot_graph}
\title{Plot a treemix graph}
\usage{
treemix_plot_graph(obj)
}
\arguments{
\item{obj}{an object of class treemixResult}
}
\value{
a ggplot object
}
\description{
Plot a treemix graph with ggplot from the treemixResult obj
}
\examples{
treemix_plot_graph('a')
}
