% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemixTools-package.R
\docType{package}
\name{treemixTools-package}
\alias{treemixTools}
\alias{treemixTools-package}
\title{treemixTools: Plot treemix result with ggplot}
\description{
Provide functionality for analysis of treemix program read and plot with ggplot
}
\keyword{internal}
